Added selected state to NumberView
